+++ semver-gradle-plugin/main/kotlin/internal/InitialVersion.kt
@@ -1,5 +1,5 @@
 package com.javiersc.semver.gradle.plugin.internal
 
-import com.javiersc.semanticVersioning.Version
+import com.javiersc.semver.Version
 
 internal val InitialVersion: Version = Version("0.1.0")